Handling empty results from region processing.
If none of the models in a dataset were present in the provided
RegionProcessor, or if region processing for any other reason would give empty
output, a ValueError would previously be raised. Now it is instead caught and
ignored, and appropriately empty results are returned by
aggregation.check_region_aggregates.

Note that the error was thrown by nomenclature, so the underlying issue is still
there. The current fix just catches the error, checks the error message against
the one thrown by nomenclature, and re-raises the exception unless the error
message matches exactly the expected one in the case of empty results. If that
error message is changed in nomenclature-iamc in the future, this fix must also
be changed.