Removed SDG indicator variables from main definitions tree.
Instead put it in a separate subfolder tree, which can be loaded as needed.
But note that some of the variables that were in the sdg.yaml file will now
need to be recreated in other files instead, and this will probably create
duplicate variables that will prevent a combined tree from loading without
raising an error. To load the SDG indicator variables alongside the other
variables, we will then probably need to write a function that loads them as
separate DataStructureDefinition or VariableCodeList instances, removes the
duplicates from one of the instances, and finally combines them into a single
instance.