Added model dimension to datastructure definition
Model name definitions have been added, so the `model` dimension can now be
included in the dsd returned by `iamcompact_nomenclature.get_dsd`.

Also changed the name of the definition file iamcompact_models.yaml to
iamcompact_iams.yaml, to reflect that we currently only include IAMs (and only
some of them), so we can include other types of models in separate files in the
future if needed.