Added region mapping file for the D4.3 harmonization data.
This is really just a map from common region names to the names of the
countries in each region. There really should be a more automated way of doing
this...

This is only the R5, R9 and R10 regions. Should still add mappings to the
prefixed model-native regions of the other models.