Commenting out R10 regions in definitions and mappings.
Turns out that the R10 regions are not unamgiguously defined in the templates
that models have used (the IPCC AR6 WGIII model registration template), which
in turn is not consistent with the definitions used in common-definitions. We
disable the R10 regions here to avoid issues when comparing harmonized data for
aggregated regions.