Restoring D4.3 region mapping file.
Will add renamed model-native regions for the IAM COMPACT models as aggregated
regions for the D4.3 data. We will temporarily require a custom branch of
nomenclature-iamc in pyproject.toml (done in the dev_main branch, to be merged)
to avoid crashes on non-ASCII characters in region names in the region maps.