Added back definitions through iamcompact-nomenclature-definitions repo.
The repo was added through a nomenclature.yaml config file in a new
subdirectory, iamcompact-nomenclature-definitions.

Note that the empty subdirectories definitions/ and mappings/ of
iamcompact-nomenclature-definitions are required for the repo to be importable
by nomenclature. The empty .gitkeep files in these directories do nothing, but
are necessary so that git will include the otherwise empty directories and not
automatically ignore them.

Note also that the iamcompact-nomenclature-definitions repository gets cloned
into yet another subdirectory during running. This subdirectory is ignored
through an added .gitignore file.