Another interim commit, need some more fixes.
Still not quite working. We need to not load RegionProcessors from more than
source after all. We are running into problems due to the fact that we are not
merging definitions and mappings from the same number of sources, and we in
fact only want to grab the region maps from *one* source anyway.