Working version with loading from external repostories, with some technical debt.
The current version takes a shortcut that creates a not quite correct `config`
attribute for the merged `DataStructureDefinition` object (it only merges the
`dimensions` and `repositories` attributes from the source DSDs, and just
copies the `definitions` and `mappings` attributes from the first one). This
is only an issue for the `.config` object, it does not appear to affect the
DataStructureDefinition object itself. The `MergedDataStructureDefinition`
object appears to need a `.config` attribute with at least a `.repositories`
field in order to be used to initialize a `RegionProcessor` object.