Created a config and directory for consolidated definitions/mappings.
This uses a single folder tree with a single nomenclature.yaml file to bring
together variables from common-definitions-fork and all other dimensions from
iamcompact-nomenclature-definitions, without having to load DSDs separately and
merge them. At the moment we don't use more than one of the repos for any one
dimension, but that should be possible if we use the filtering functionality in
the most recent versions of nomenclature.